% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_formatting_functions.R
\name{add_RS_to_GWAS}
\alias{add_RS_to_GWAS}
\title{Find overlapping RS ids in GWAS using genomic co-ordinates.}
\usage{
add_RS_to_GWAS(GWAS, dbSNPref)
}
\arguments{
\item{GWAS}{GWAS with a minimum of 2 columns labelled CHR and BP. If
additional columns included these will be preserved.}

\item{dbSNPref}{BS genome reference snps (choose appropriate dbSNP build
dependent on genome build)}
}
\value{
GWAS with RS ids.
}
\description{
Function to find the RS ids in a GWAS where SNPs have been provided as
genomic coordinates i.e. some combination of SNP and BP.
}
