% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{AddBPWindow}
\alias{AddBPWindow}
\title{Add window to genes}
\usage{
AddBPWindow(dataset, windowsize, mouse = NULL)
}
\arguments{
\item{dataset}{Dataframe with columns with gene start_position and
end_position. Check columns names are start_position and end_position for
human queries.}

\item{windowsize}{Size of window to be added to annotation}

\item{mouse}{If original genes are from mouse then select "true". Necessary
as output column names from biomart queries differs between species.}
}
\value{
Dataframe with x bp added to gene start/end.
}
\description{
Function for adding a window to genes.
}
