% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{AddGenePosDetails_RemoveXandYandMT}
\alias{AddGenePosDetails_RemoveXandYandMT}
\title{Function for extracting human gene positional information and removing X, Y
and MT chromosomal genes.}
\usage{
AddGenePosDetails_RemoveXandYandMT(dataframe, columnToFilter, mart = 38,
  attributes, filter)
}
\arguments{
\item{dataframe}{Dataframe with gene names.}

\item{columnToFilter}{Name of column in dataframe, which contains gene names.}

\item{mart}{Specify genome build. Baseline from LDSC based on hg19/37.}

\item{attributes}{Vector of attributes to extract from BioMart.}

\item{filter}{Vector of filter to be used for BioMart query.}
}
\value{
Original dataframe together with gene positional information.
}
\description{
Function for extracting human gene positional information and removing X, Y
and MT chromosomal genes.
}
