% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{Assimilate.H2.results.multipleannot}
\alias{Assimilate.H2.results.multipleannot}
\title{Function assembling heritability results from LD score regression run with multiple annotations.}
\usage{
Assimilate.H2.results.multipleannot(path_to_results)
}
\arguments{
\item{path_to_results}{ist of file paths where .results output from LDSC stored}
}
\value{
Dataframe of results together with associated annotations and GWAS.
}
\description{
Function assembling heritability results from LD score regression run with multiple annotations.
}
