% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{Assimilate_H2_results}
\alias{Assimilate_H2_results}
\title{Function assembling herirtability results from LD score regression}
\usage{
Assimilate_H2_results(path_to_results)
}
\arguments{
\item{path_to_results}{List of file paths where .results output from LDSC
stored}
}
\value{
Dataframe of results together with annotation name and GWAS.
}
\description{
Function assembling herirtability results from LD score regression
}
