% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{Calculate_enrichment_SE_and_logP}
\alias{Calculate_enrichment_SE_and_logP}
\title{Calculate enrichment SE, logP and z-score p-value.}
\usage{
Calculate_enrichment_SE_and_logP(df, one_sided)
}
\arguments{
\item{df}{Dataframe with .results output of LDSC, as created using the
Assimilate_H2_results() function.}

\item{one_sided}{Specify whether one-sided or two-sided p-value required. If
NULL, two-sided test will be used. If one-sided preferred then specify
direction of one-sided test. "-" for negative, and "+" for positive.}
}
\value{
Dataframe with upper and lower bounds of enrichment SE, log
  P(enrichment) and z-score p-value
}
\description{
Function calculating upper and lower bounds of enrichment SE, log
P(enrichment) and z-score p-value.
}
