% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{convert_z_score}
\alias{convert_z_score}
\title{Convert z-score to p-value}
\usage{
convert_z_score(z, one_sided = NULL)
}
\arguments{
\item{z}{Z-score}

\item{one_sided}{Specify whether one-sided or two-sided p-value required.
If NULL, two-sided test will be used. If one-sided preferred then specify direction of one-sided test.
"-" for negative, and "+" for positive.}
}
\value{
P-value.
}
\description{
Function converting z-score to p-value.
}
