% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{create_annot_file_and_export}
\alias{create_annot_file_and_export}
\title{Function to create 22 .annot.gz files for each annotation.}
\usage{
create_annot_file_and_export(list_of_annotations, annot_basedir)
}
\arguments{
\item{list_of_annotations}{A list with all annotations that have been
overlapped with the baseline model.}

\item{annot_basedir}{Base directory in which files should be created. Do not
include "/" at the end of the directory name.}
}
\value{
Create directories representing each of the annotations, with
  .annot.gz files for each of 22 chromosomes.
}
\description{
Function to create 22 .annot.gz files for each annotation.
}
