% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Pipeline_Functions.R
\name{Create_GWAS_df}
\alias{Create_GWAS_df}
\title{Create GWAS dataframe.}
\usage{
Create_GWAS_df()
}
\value{
Dataframe of availabe GWASs in the LDscore GWAS directory.
}
\description{
Function that creates a GWAS dataframe based on all of the .sumstats.gz files available in the LDscore GWAS directory.
}
