% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{df2GRanges}
\alias{df2GRanges}
\title{Convert from dataframe to GRanges}
\usage{
df2GRanges(df, seqname_col, start_col, end_col)
}
\arguments{
\item{df}{Dataframe with columns as sequence name, start, and end. Extra
columns kept as metadata.}

\item{seqname_col}{Name of column containing chromsome name. Must be in
quotation marks.}

\item{start_col}{Name of column containing start position. Must be in
quotation marks.}

\item{end_col}{Name of column containing end position. Must be in quotation
marks.}
}
\value{
GRanges object with original data from dataframe.
}
\description{
Convert from dataframe to GRanges
}
