% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Pipeline_Functions.R
\name{get_LDSC_fixed_args}
\alias{get_LDSC_fixed_args}
\title{Get fixed LDSC arguments.}
\usage{
get_LDSC_fixed_args(Baseline_model = NULL)
}
\arguments{
\item{Baseline_model}{Choose baseline model to run heritability estimates
with. Choose by specifying number of annotations: 53, 75, 76, 86, or 97.}
}
\value{
List of fixed LDSC arguments.
}
\description{
Function to return a list of the fixed arguments. To be used in combination
with the Calculate_LDscore() and Calculate_H2() functions.
}
