% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{keep_duplicates}
\alias{keep_duplicates}
\title{Keep duplicates}
\usage{
keep_duplicates(Dataframe, Column_w_duplicates)
}
\arguments{
\item{Dataframe}{Dataframe with a column containing duplicated values.}

\item{Column_w_duplicates}{Column name in quotation marks.}
}
\value{
Dataframe with only duplicated rows.
}
\description{
Function for extracting duplicated values in a defined column.
}
