% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_formatting_functions.R
\name{liftover_hg19_to_hg38}
\alias{liftover_hg19_to_hg38}
\title{Liftover GWAS from hg19 to hg38}
\usage{
liftover_hg19_to_hg38(GWAS, path_to_chain)
}
\arguments{
\item{GWAS}{GWAS with a minimum of 2 columns CHR and BP. If
additional columns included these will be preserved.}

\item{path_to_chain}{Path to UCSC chain file for transformation from hg19 to
hg38 coordinates.}
}
\value{
GWAS with hg38 coordinates
}
\description{
Liftover GWAS from hg19 to hg38
}
