% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{overlap_annot_hits_w_baseline}
\alias{overlap_annot_hits_w_baseline}
\title{Map annotation SNPs overlapping the baseline model back to the full baseline
model.}
\usage{
overlap_annot_hits_w_baseline(list_of_annotations, BM)
}
\arguments{
\item{list_of_annotations}{List of annotation dataframes containing only
those SNPs that were found to overlap the baseline model using the
overlap_annot_list() function.}

\item{BM}{Baseline model with columns: CHR, BP, SNP, CM.}
}
\value{
List of annotations with all SNPs present in the baseline model, in
  addition to a column distinguishing between annotation SNPs present in the
  baseline model (1) and those not found within the baseline model (0).
}
\description{
Function is used to map annotation SNPs overlapping the baseline model back
to the full baseline model, with a column of 1s and 0s to distinguish between
SNPs overlapping the baseline model (1) and those that do not overlap (0).
Function is intended for use after using the overlap_annot_list() function.
}
