% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{Plot.H2.coefficient}
\alias{Plot.H2.coefficient}
\title{Function plotting coefficient p-value of annotation, with GWASs facetted.}
\usage{
Plot.H2.coefficient(h2.results, x.axis, xlab, fill.variable, colour,
  pvalue.cutoff, show.legend)
}
\arguments{
\item{h2.results}{Dataframe with overall heritability.}

\item{x.axis}{What variable should be plotted on the x-axis.}

\item{xlab}{X-axis name in quotation marks.}

\item{fill.variable}{What variable should be used to determine the fill
colour of bars/points.}

\item{colour}{Vector with colours.}

\item{pvalue.cutoff}{P-value cutoff after multiple test correction for number
of annotations and GWASs.}

\item{show.legend}{TRUE or FALSE depending on it want legend shown}
}
\value{
Plot
}
\description{
Function plotting coefficient p-value of annotation, with GWASs facetted.
}
