% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{Plot_H2_enrichment_coefficient}
\alias{Plot_H2_enrichment_coefficient}
\title{Plot S-LDSC H2 enrichment & coefficient per annotation}
\usage{
Plot_H2_enrichment_coefficient(
  h2_results,
  x_axis,
  xlab,
  fill_variable = NULL,
  colour,
  pvalue_cutoff
)
}
\arguments{
\item{h2_results}{Dataframe with overall heritability.}

\item{x_axis}{character. Variable to be plotted on the x-axis.}

\item{xlab}{character. X-axis name in quotation marks.}

\item{fill_variable}{character. Variable to be used to determine the fill
colour of bars/points. Default is NULL.}

\item{colour}{character vector. If fill_variable not NULL then supply colours
to be used.}

\item{pvalue_cutoff}{numeric. P-value cutoff after multiple test correction
for number of annotations and GWASs.}
}
\value{
Plot
}
\description{
Function plotting enrichment, coefficient and coefficient p-value of
annotation, with GWASs facetted.
}
