% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates}
\usage{
remove_duplicates(Dataframe, Column_w_duplicates)
}
\arguments{
\item{Dataframe}{Dataframe with columns as gene name and enrichment values.}

\item{Column_w_duplicates}{Column name in quotation marks.}
}
\value{
Original dataframe without duplicated rows.
}
\description{
Function for removing duplicates within defined column.
}
