% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_Creatingannot_Functions.R
\name{create_annot_file_and_export}
\alias{create_annot_file_and_export}
\title{Function to create 22 .annot.gz files for each annotation.}
\usage{
create_annot_file_and_export(list_of_annotations, annot_basedir)
}
\arguments{
\item{list_of_annotations}{A named list with all annotations that have been
overlapped with the baseline model. It is important that annotation names
do not include an underscore ('_'). If it does, when using the
Assimilate_H2_results() function anything prior to the last word in the
annotation name will be removed, due to the way the Assimilate_H2_results()
function works. Consider using a different separate, such as a full stop or
colon.}

\item{annot_basedir}{Base directory in which files should be created. Do not
include "/" at the end of the directory name.}
}
\value{
Create directories representing each of the named annotations, with
  .annot.gz files for each of 22 chromosomes.
}
\description{
Function to create 22 .annot.gz files for each annotation.
}
