% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDSC_SummariseOutput_Functions.R
\name{Plot_H2_enrichment_coefficient}
\alias{Plot_H2_enrichment_coefficient}
\title{Plot S-LDSC H2 enrichment & coefficient per annotation}
\usage{
Plot_H2_enrichment_coefficient(h2.results, x.axis, xlab, fill.variable,
  colour, pvalue.cutoff)
}
\arguments{
\item{h2.results}{Dataframe with overall heritability.}

\item{x.axis}{What variable should be plotted on the x-axis.}

\item{xlab}{X-axis name in quotation marks.}

\item{fill.variable}{What variable should be used to determine the fill
colour of bars/points.}

\item{colour}{Vector with colours.}

\item{pvalue.cutoff}{P-value cutoff after multiple test correction for number
of annotations and GWASs.}
}
\value{
Plot
}
\description{
Function plotting enrichment, coefficient and coefficient p-value of
annotation, with GWASs facetted.
}
